Nearly finished comments table functionality. The only thing missing are dynamically resized tableViewCells based around the size of comment text.I also implemented the new string photo implementation. You can manually take/choose photos and insert them into a string.

+++ Stringr/Detail/Edit/StringrStringDetailEditTableViewController.m
@@ -176,7 +176,7 @@
 
 #pragma mark - TableView Delegate
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     if (indexPath.section == 0 && indexPath.row == 2) {
         return 110;
